+++ packages/shared/src/utils/index.ts
@@ -10,6 +10,7 @@ export * from "./core/dateUtils";
 export * from "./core/numberUtils";
 export * from "./export.utils";
 export * from "./compositionDocumentOrder";
+export * from "./disclosureGroupExpansion";
 export * from "./migrateCollectionItems";
 // ADR-076: BC re-export (deprecated, 신규 코드는 migrateCollectionItems 사용)
 export {